+++ tortoise/backends/asyncpg/schema_generator.py
@@ -10,7 +10,7 @@ if TYPE_CHECKING:  # pragma: nocoverage
 class AsyncpgSchemaGenerator(BaseSchemaGenerator):
     DIALECT = "postgres"
     TABLE_COMMENT_TEMPLATE = "COMMENT ON TABLE \"{table}\" IS '{comment}';"
-    COLUMN_COMMNET_TEMPLATE = 'COMMENT ON COLUMN "{table}"."{column}" IS \'{comment}\';'
+    COLUMN_COMMENT_TEMPLATE = 'COMMENT ON COLUMN "{table}"."{column}" IS \'{comment}\';'
     GENERATED_PK_TEMPLATE = '"{field_name}" {generated_sql}'
 
     def __init__(self, client: "AsyncpgDBClient") -> None:
@@ -31,7 +31,7 @@ class AsyncpgSchemaGenerator(BaseSchemaGenerator):
         return ""
 
     def _column_comment_generator(self, table: str, column: str, comment: str) -> str:
-        comment = self.COLUMN_COMMNET_TEMPLATE.format(
+        comment = self.COLUMN_COMMENT_TEMPLATE.format(
             table=table, column=column, comment=self._escape_comment(comment)
         )
         if comment not in self.comments_array: