+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 - Add more `index` types.
 - Add `force_index`, `use_index` to `queryset`.
 - Fix `F` in update error with `update_fields`.
+- Make `delete` query work with `limit` and `order_by`. (#697)
 
 0.17.1
 ------